requirements.txt
fastapi
uvicorn
motor
beanie
pymongo
pydantic[email]
scikit-learn
pandas
numpy
requests
python-dotenv
gunicorn
passlib[bcrypt]
python-jose[cryptography]
python-multipart
openpyxl
razorpay
email-validator
